adding buttons

+++ tower/src/explosive_tower.cpp
@@ -7,7 +7,7 @@ ExplosiveTower::ExplosiveTower(Vector2 position, const ResourceManager& RM)
     m_damage = 10;
     m_cooldown = m_fire_rate = 2.5f;
     m_range = 250;
-    m_texture = RM.get_texture((unsigned int)TextureType::TOWER_IDLE);
+    m_texture = RM.get_texture(TextureType::TOWER_IDLE);
 }
 
 void ExplosiveTower::update()
