+++ tower/src/healer_enemy.cpp
@@ -5,7 +5,7 @@ HealerEnemy::HealerEnemy(Vector2 position, const ResourceManager& RM) : Enemy(po
     m_life = Life(25);
     m_damage = 5;
     m_speed = 400;
-    m_texture = RM.get_texture((unsigned int)TextureType::ENEMY_IDLE);
+    m_texture = RM.get_texture(TextureType::ENEMY_IDLE);
 }
 
 void HealerEnemy::update()
